cleaned up addfriend
